requirements.txt
asgiref==3.5.1
Django==3.2
django-ckeditor==6.3.2
django-debug-toolbar==3.4.0
django-js-asset==2.0.0
django-mptt==0.13.4
flake8==4.0.1
importlib-metadata==4.2.0
mccabe==0.6.1
Pillow==9.1.0
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
zipp==3.8.0
